\name{QuantifyPeaks}
\alias{QuantifyPeaks}
\title{Function to quantify the average peak profiles of ChIPseq data}
\usage{
hello()
}
\description{
Prints 'Hello, world!'.
}
\examples{
hello()
}


\name{QuantifyPeaks}
\alias{QuantifyPeaks}
\title{Function to quantify the average peak profiles of ChIPseq data}
\usage{
hello()
}
\description{
Prints 'Hello, world!'.
}
\examples{
hello()
}
